+++ iaso/management/commands/delete_accounts.py
@@ -39,7 +39,8 @@ from iaso.models.device import Device
 from iaso.models.entity import Entity, EntityType
 from iaso.models.forms import Form
 from iaso.models.instances import Instance, InstanceFile, InstanceLock
-from iaso.models.microplanning import Assignment, Planning, Team
+from iaso.models.microplanning import Assignment, Planning
+from iaso.models.team import Team
 from iaso.models.org_unit import OrgUnit, OrgUnitReferenceInstance
 from iaso.models.pages import Page
 from iaso.models.project import Project
