+++ iaso/management/commands/seed_test_data.py
@@ -41,7 +41,8 @@ from iaso.models.base import AccountFeatureFlag
 from iaso.models.comment import CommentIaso
 from iaso.models.device import Device
 from iaso.models.entity import Entity, EntityType
-from iaso.models.microplanning import Planning, Team
+from iaso.models.microplanning import Planning
+from iaso.models.team import Team
 from iaso.models.pages import Page
