+++ iaso/models/__init__.py
@@ -10,7 +10,8 @@ from .forms import Form, FormAttachment, FormPredefinedFilter, FormVersion
 from .import_gpkg import ImportGPKG
 from .instances import Instance, InstanceFile, InstanceLock, InstanceQuerySet
 from .metric import MetricType, MetricValue
-from .microplanning import Planning, Team
+from .microplanning import Planning
+from .team import Team
 from .openhexa import OpenHEXAInstance, OpenHEXAWorkspace
 from .org_unit import OrgUnit, OrgUnitChangeRequest, OrgUnitReferenceInstance, OrgUnitType
 from .org_unit_change_request_configuration import OrgUnitChangeRequestConfiguration
