+++ iaso/models/microplanning.py
@@ -1,124 +1,11 @@
-import operator
-import typing
-
-from functools import reduce
-
 from django.contrib.auth.models import User
 from django.contrib.postgres.fields import ArrayField
-from django.db import models, transaction
+from django.db import models
 from django.db.models import Q
-from django_ltree.fields import PathField  # type: ignore
 
 from iaso.models import Form, OrgUnit, Project
-from iaso.utils.expressions import ArraySubquery
+from iaso.models.team import Team
 from iaso.utils.models.soft_deletable import SoftDeletableModel
-
-
-class TeamQuerySet(models.QuerySet):
-    def filter_for_user(self, user: User):
-        return self.filter(project__account=user.iaso_profile.account)
-
-    def hierarchy(self, teams: typing.Union[typing.List["Team"], "models.QuerySet[Team]", "Team"]) -> "TeamQuerySet":
-        """The Team and all their descendants"""
-        if isinstance(teams, Team):
-            query = models.Q(path__descendants=teams.path)
-        elif isinstance(teams, models.QuerySet):
-            team_qs = teams
-            query = models.Q(path__descendants=ArraySubquery(team_qs.values("path")))
-        elif isinstance(teams, (list,)):
-            query = reduce(operator.or_, [models.Q(path__descendants=str(ou.path)) for ou in teams])
-        else:
-            raise NotImplementedError
-
-        return self.filter(query)
-
-
-class TeamType(models.TextChoices):
-    TEAM_OF_TEAMS = "TEAM_OF_TEAMS", "Team of teams"
-    TEAM_OF_USERS = "TEAM_OF_USERS", "Team of users"
-
-    @classmethod
-    def is_valid_team_type(cls, value):
-        return value in cls.values
-
-
-TeamManager = models.Manager.from_queryset(TeamQuerySet)
-
-
-class Team(SoftDeletableModel):
-    objects = TeamManager()
-
-    class Meta:
-        ordering = ("name",)
-
-    name = models.CharField(max_length=100, null=False, blank=False)
-    description = models.TextField(blank=True)
-    project = models.ForeignKey(Project, on_delete=models.PROTECT)
-    users = models.ManyToManyField(User, related_name="teams", blank=True)
-    manager = models.ForeignKey(User, on_delete=models.PROTECT, related_name="managed_teams")
-    parent = models.ForeignKey("self", on_delete=models.PROTECT, null=True, blank=True, related_name="sub_teams")
-    path = PathField(unique=True)
-    # scope = models.ManyToManyField("OrgUnit", related_name="teams")
-    type = models.CharField(choices=TeamType.choices, max_length=100, null=True, blank=True)
-
-    created_by = models.ForeignKey(User, on_delete=models.PROTECT, null=True, blank=True)
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-
-    # Path management logic, taken from org_unit.
-    def save(self, *args, skip_calculate_path: bool = False, force_recalculate: bool = False, **kwargs):
-        """Override default save() to make sure that the path property is calculated and saved,
-        for this org unit and its children.
-
-        :param skip_calculate_path: use with caution - can be useful in scripts where the extra transactions
-                                    would be a burden, but the path needs to be set afterwards
-        :param force_recalculate: use with caution - used to force recalculation of paths
-        """
-        # correct type for parent type
-        if self.parent and self.parent.type is None:
-            self.parent.type = TeamType.TEAM_OF_TEAMS
-            self.parent.save(skip_calculate_path=True)
-
-        if skip_calculate_path:
-            super().save(*args, **kwargs)
-        else:
-            with transaction.atomic():
-                super().save(*args, **kwargs)
-                Team.objects.bulk_update(self.calculate_paths(force_recalculate=force_recalculate), ["path"])
-
-    def calculate_paths(self, force_recalculate: bool = False) -> typing.List["Team"]:
-        """Calculate the path for this Team and all its children recursively.
-
-        This method will check if this instance path should change. If it is the case (or if force_recalculate is
-        True), it will update the path property for the instance and its children, and return all the modified
-        records.
-
-        Please note that this method does not save the modified records. Instead, they are updated in bulk in the
-        save() method.
-
-        :param force_recalculate: calculate path for all descendants, even if this instance path does not change
-        """
-
-        # keep track of updated records
-        updated_records = []
-
-        # noinspection PyUnresolvedReferences
-        base_path = [] if self.parent is None else list(self.parent.path)
-        new_path = [*base_path, str(self.pk)]
-        path_has_changed = new_path != self.path
-
-        if path_has_changed:
-            self.path = new_path
-            updated_records += [self]
-
-        if path_has_changed or force_recalculate:
-            for child in self.sub_teams.all():
-                updated_records += child.calculate_paths(force_recalculate)
-
-        return updated_records
-
-    def __str__(self):
-        return self.name
 
 
 class PlanningQuerySet(models.QuerySet):
