+++ iaso/tests/api/test_microplanning.py
@@ -10,7 +10,8 @@ from django_ltree.fields import PathValue  # type: ignore
 from hat.audit.models import Modification
 from iaso.api.microplanning.serializers import AssignmentSerializer, PlanningSerializer, TeamSerializer
 from iaso.models import Account, DataSource, Form, OrgUnit, OrgUnitType, SourceVersion
-from iaso.models.microplanning import Assignment, Planning, Team, TeamType
+from iaso.models.microplanning import Assignment, Planning
+from iaso.models.team import Team, TeamType
 from iaso.permissions.core_permissions import CORE_PLANNING_WRITE_PERMISSION, CORE_TEAMS_PERMISSION
 from iaso.test import APITestCase, IasoTestCaseMixin
