+++ plugins/polio/models/base.py
@@ -35,7 +35,7 @@ from iaso.models import Group, OrgUnit
 from iaso.models.base import Account
 from iaso.models.entity import UserNotAuthError
 from iaso.models.json_config import Config
-from iaso.models.microplanning import Team
+from iaso.models.team import Team
 from iaso.models.project import Project
 from iaso.models.task import Task
 from iaso.utils import slugify_underscore
